kv: move test_utils file to new kvtestutils package
Release note: None
