+++ pkg/BUILD.bazel
@@ -1586,6 +1586,7 @@ GO_TARGETS = [
     "//pkg/kv/kvserver/uncertainty:uncertainty_test",
     "//pkg/kv/kvserver:kvserver",
     "//pkg/kv/kvserver:kvserver_test",
+    "//pkg/kv/kvtestutils:kvtestutils",
     "//pkg/kv:kv",
     "//pkg/kv:kv_test",
     "//pkg/multitenant/mtinfo:mtinfo",
