+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -242,8 +242,8 @@ go_library(
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
         "//pkg/keys",
-        "//pkg/kv",
         "//pkg/kv/kvpb",
+        "//pkg/kv/kvtestutils",
         "//pkg/multitenant/mtinfopb",
         "//pkg/roachpb",
         "//pkg/roachprod",
