+++ pkg/kv/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "mock_transactional_sender.go",
         "range_lookup.go",
         "sender.go",
-        "test_utils.go",
         "txn.go",
         "util.go",
     ],
@@ -19,7 +18,6 @@ go_library(
     deps = [
         "//pkg/base",
         "//pkg/keys",
-        "//pkg/kv/kvbase",
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver/closedts",
         "//pkg/kv/kvserver/concurrency/isolation",
@@ -28,7 +26,6 @@ go_library(
         "//pkg/settings/cluster",
         "//pkg/sql/sessiondatapb",
         "//pkg/storage/enginepb",
-        "//pkg/testutils",
         "//pkg/util/admission",
         "//pkg/util/admission/admissionpb",
         "//pkg/util/duration",
@@ -40,8 +37,6 @@ go_library(
         "//pkg/util/stop",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
-        "//pkg/util/tracing",
-        "//pkg/util/tracing/tracingpb",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_errors//:errors",
@@ -75,6 +70,7 @@ go_test(
         "//pkg/kv/kvserver/concurrency/lock",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/kv/kvserver/liveness/livenesspb",
+        "//pkg/kv/kvtestutils",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
