+++ pkg/kv/kvtestutils/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "kvtestutils",
+    srcs = ["test_utils.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvtestutils",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/kv/kvbase",
+        "//pkg/testutils",
+        "//pkg/util/tracing",
+        "//pkg/util/tracing/tracingpb",
+    ],
+)
